+++ src/main/java/de/samply/generator/TestDataGenerator.java
@@ -1,0 +1,15 @@
+package de.samply.generator;
+
+import de.samply.model.OBDS;
+import org.springframework.stereotype.Component;
+
+@Component
+public class TestDataGenerator {
+
+    public OBDS generate() {
+        OBDS obds = new OBDS();
+        //TODO
+        return obds;
+    }
+
+}